pyproject.toml
[project]
name = "researchpapers"
version = "0.1.0"
description = ""
authors = [
    {name = "Aliveni1702",email = "alivenimangalapati@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "lxml (>=5.3.1,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
